+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog keeps track of all changes to the Packs SDK. We follow convention
 
 ## [Unreleased]
 
+### Changed
+
+- Add `validateParameters` metadata formula to CommonPackFormulaDef
+
 ## [1.9.5] - 2025-05-02
 
 ### Changed